+++ sdk/lib/js_interop/js_interop.dart
@@ -26,6 +26,7 @@ import 'dart:typed_data';
 /// Allow use of `@staticInterop` classes with JS types as well as export
 /// functionality.
 export 'dart:_js_annotations' show staticInterop, anonymous, JSExport;
+export 'dart:js_util' show NullRejectionException;
 
 /// The annotation for JS interop members.
 ///
@@ -247,7 +248,22 @@ extension NullableUndefineableJSAnyExtension on JSAny? {
 
   bool get isUndefinedOrNull => this == null;
   bool get isDefinedAndNotNull => !isUndefinedOrNull;
+}
+
+/// Common utility functions that are useful for any JS value.
+extension JSAnyUtilityExtension on JSAny? {
+  /// Returns whether the result of `typeof` on this [JSAny]? is [typeString].
   external bool typeofEquals(String typeString);
+
+  /// Returns whether this [JSAny]? is an `instanceof` [constructor].
+  external bool instanceof(JSFunction constructor);
+
+  /// Like [instanceof], but only takes a [String] for the constructor name,
+  /// which is then looked up in the [globalContext].
+  bool instanceOfString(String constructorName) {
+    final constructor = globalContext[constructorName] as JSFunction?;
+    return constructor != null && instanceof(constructor);
+  }
 
   /// Effectively the inverse of [jsify], [dartify] Takes a JavaScript object,
   /// and converts it to a Dart based object. Only JS primitives, arrays, or
@@ -260,18 +276,6 @@ extension NullableObjectUtilExtension on Object? {
   /// Recursively converts a JSON-like collection, or Dart primitive to a
   /// JavaScript compatible representation.
   external JSAny? jsify();
-}
-
-/// Utility extensions for [JSObject].
-extension JSObjectUtilExtension on JSObject {
-  external bool instanceof(JSFunction constructor);
-
-  /// Like [instanceof], but only takes a [String] for the constructor name,
-  /// which is then looked up in the [globalContext].
-  bool instanceOfString(String constructorName) {
-    final constructor = globalContext[constructorName] as JSFunction?;
-    return constructor != null && instanceof(constructor);
-  }
 }
 
 /// The type of `JSUndefined` when returned from functions. Unlike pure JS,
@@ -544,3 +548,122 @@ extension JSStringToString on JSString {
 extension StringToJSString on String {
   external JSString get toJS;
 }
+
+// General-purpose operators.
+//
+// Indexing operators (`[]`, `[]=`) should be declared through operator
+// overloading instead e.g. `external operator int [](int key);`.
+// TODO(srujzs): Add more as needed. For now, we just expose the ones needed to
+// migrate from `dart:js_util`.
+extension JSAnyOperatorExtension on JSAny? {
+  // Artithmetic operators.
+
+  /// Returns the result of '[this] + [any]' in JS.
+  external JSAny add(JSAny? any);
+
+  /// Returns the result of '[this] - [any]' in JS.
+  external JSAny subtract(JSAny? any);
+
+  /// Returns the result of '[this] * [any]' in JS.
+  external JSAny multiply(JSAny? any);
+
+  /// Returns the result of '[this] / [any]' in JS.
+  external JSAny divide(JSAny? any);
+
+  /// Returns the result of '[this] % [any]' in JS.
+  external JSAny modulo(JSAny? any);
+
+  /// Returns the result of '[this] ** [any]' in JS.
+  external JSAny exponentiate(JSAny? any);
+
+  // Comparison operators.
+
+  /// Returns the result of '[this] > [any]' in JS.
+  external JSBoolean greaterThan(JSAny? any);
+
+  /// Returns the result of '[this] >= [any]' in JS.
+  external JSBoolean greaterThanOrEqualTo(JSAny? any);
+
+  /// Returns the result of '[this] < [any]' in JS.
+  external JSBoolean lessThan(JSAny? any);
+
+  /// Returns the result of '[this] <= [any]' in JS.
+  external JSBoolean lessThanOrEqualTo(JSAny? any);
+
+  /// Returns the result of '[this] == [any]' in JS.
+  external JSBoolean equals(JSAny? any);
+
+  /// Returns the result of '[this] != [any]' in JS.
+  external JSBoolean notEquals(JSAny? any);
+
+  /// Returns the result of '[this] === [any]' in JS.
+  external JSBoolean strictEquals(JSAny? any);
+
+  /// Returns the result of '[this] !== [any]' in JS.
+  external JSBoolean strictNotEquals(JSAny? any);
+
+  // Bitwise operators.
+
+  /// Returns the result of '[this] >>> [any]' in JS.
+  external JSNumber unsignedRightShift(JSAny? any);
+
+  // Logical operators.
+
+  /// Returns the result of '[this] && [any]' in JS.
+  external JSAny? and(JSAny? any);
+
+  /// Returns the result of '[this] || [any]' in JS.
+  external JSAny? or(JSAny? any);
+
+  /// Returns the result of '![this]' in JS.
+  external bool get not;
+
+  /// Returns the result of '!![this]' in JS.
+  external bool get isTruthy;
+}
+
+/// Given a Dart object that is marked "exportable", creates a JS object that
+/// wraps the given Dart object. Look at the `@JSExport` annotation to determine
+/// what constitutes "exportable" for a Dart class. The object literal
+/// will be a map of export names (which are either the written instance member
+/// names or their rename) to their respective Dart instance members.
+///
+/// For example:
+///
+/// ```
+/// import 'dart:js_interop';
+///
+/// import 'package:expect/expect.dart';
+///
+/// @JSExport()
+/// class ExportCounter {
+///   @JSExport('value')
+///   int counterValue = 0;
+///   String stringify() => counterValue.toString();
+/// }
+///
+/// extension type Counter(JSObject _) {
+///   external int get value;
+///   external set value(int val);
+///   external String stringify();
+/// }
+///
+/// void main() {
+///   var export = ExportCounter();
+///   var counter = Counter(createJSInteropWrapper(export));
+///   export.counterValue = 1;
+///   Expect.equals(counter.value, export.counterValue);
+///   Expect.equals(counter.stringify(), export.stringify());
+/// }
+/// ```
+external JSObject createJSInteropWrapper<T extends Object>(T dartObject);
+
+// TODO(srujzs): Expose this method when we handle conformance checking for
+// interop extension types. We don't expose this method today due to the bound
+// on `T`. `@staticInterop` types can't implement `JSObject`, so this method
+// simply wouldn't work. We could make it extend `Object` to support the
+// `@staticInterop` case, but if we ever refactor to `extends JSObject`, this
+// would be a breaking change. For now, due to the low usage of
+// `createStaticInteropMock`, we avoid introducing this method until later.
+// external T createJSInteropMock<T extends JSObject, U extends Object>(
+//     U dartMock, [JSObject? proto = null]);