+++ tests/lib/js/static_interop_test/js_types_test.dart
@@ -6,7 +6,6 @@
 
 import 'dart:js_interop';
 import 'dart:js_interop_unsafe';
-import 'dart:js_util';
 import 'dart:typed_data';
 
 import 'package:expect/expect.dart';