+++ tests/lib/js_interop_unsafe/basic_test.dart
@@ -133,8 +133,10 @@ void instanceOfTest() {
   JSFunction jsClass2Constructor = gc['JSClass2'] as JSFunction;
   Expect.isTrue(obj.instanceof(jsClass1Constructor));
   Expect.isFalse(obj.instanceof(jsClass2Constructor));
+  Expect.isFalse(0.toJS.instanceof(jsClass2Constructor));
   Expect.isTrue(obj.instanceOfString('JSClass1'));
   Expect.isFalse(obj.instanceOfString('JSClass2'));
+  Expect.isFalse(0.toJS.instanceOfString('JSClass1'));
 }
 
 void _expectIterableEquals(Iterable<Object?> l, Iterable<Object?> r) {